+++ source/renderer/app/themes/daedalus/yellow.js
@@ -886,25 +886,35 @@ export const YELLOW_THEME_OUTPUT = {
       'rgba(234, 76, 91, .1)',
   },
   walletImportDialog: {
-    '--theme-wallet-import-background-color': '#fdcd68f5',
+    '--theme-wallet-import-background-color': 'rgba(255, 185, 35, 0.98)',
+    '--theme-wallet-import-button-background-color': 'rgba(45, 45, 45, 0.1)',
+    '--theme-wallet-import-button-border-color': '#2d2d2d',
+    '--theme-wallet-import-button-text-color': '#2d2d2d',
+    '--theme-wallet-import-button-text-color-hover': '#fff',
+    '--theme-wallet-import-checkbox-border': '2px solid #ffffff',
+    '--theme-wallet-import-checkbox-border-color': '#2d2d2d',
+    '--theme-wallet-import-checkbox-border-color-disabled': '#fff',
+    '--theme-wallet-import-checkbox-check-bg-color': '#2d2d2d',
+    '--theme-wallet-import-checkbox-check-color': '#ffb923',
+    '--theme-wallet-import-description-close-color': '#2d2d2d',
     '--theme-wallet-import-description-close-hover-background':
       'rgba(0, 0, 0, 0.1)',
-    '--theme-wallet-import-description-color': 'rgba(255, 255, 255, 0.7)',
-    '--theme-wallet-import-description-bold-color': '#fff',
-    '--theme-wallet-import-title-color': '#fff',
-    '--theme-wallet-import-stateFolder-border-color':
-      'rgba(255, 255, 255, 0.5)',
+    '--theme-wallet-import-description-color': 'rgba(45, 45, 45, 0.7)',
+    '--theme-wallet-import-description-bold-color': '#2d2d2d',
+    '--theme-wallet-import-error-color': '#be0b0b',
+    '--theme-wallet-import-input-background-color': 'rgba(0, 0, 0, 0.1)',
+    '--theme-wallet-import-input-background-color-disabled':
+      'rgba(45, 45, 45, 0.05)',
+    '--theme-wallet-import-input-border-color': 'rgba(45, 45, 45, 0.3)',
+    '--theme-wallet-import-stateFolder-border-color': 'rgba(45, 45, 45, 0.3)',
     '--theme-wallet-import-stateFolder-button-background-color':
-      'rgba(255, 255, 255, 0.1)',
+      'rgba(45, 45, 45, 0.1)',
     '--theme-wallet-import-stateFolder-button-background-color-hover':
-      'rgba(255, 255, 255, 0.05)',
+      'rgba(45, 45, 45, 0.05)',
     '--theme-wallet-import-stateFolder-button-background-color-active':
-      'rgba(255, 255, 255, 0.12)',
-    '--theme-wallet-import-checkbox-border': '2px solid #ffffff',
-    '--theme-wallet-import-checkbox-border-color': 'rgba(255, 255, 255, 1)',
-    '--theme-wallet-import-checkbox-border-color-disabled': '#ffffff',
-    '--theme-wallet-import-checkbox-check-bg-color': '#ffffff',
-    '--theme-wallet-import-stateFolder-label-color': '#fff',
+      'rgba(45, 45, 45, 0.12)',
+    '--theme-wallet-import-stateFolder-label-color': '#2d2d2d',
+    '--theme-wallet-import-title-color': '#2d2d2d',
   },
   walletRestoreDialog: {
     '--theme-wallet-restore-dialog-new-label-background-color':